+++ FTMobileSDK/FTMobileAgent/Config/FTMobileConfig+Private.h
@@ -7,7 +7,7 @@
 //
 
 #import "FTMobileConfig.h"
-#import "FTLoggerConfig.h"
+#import "FTLoggerConfig+Private.h"
 NS_ASSUME_NONNULL_BEGIN
 @interface FTMobileConfig ()
 /// 添加 pkg 信息
@@ -22,14 +22,9 @@ NS_ASSUME_NONNULL_BEGIN
 -(instancetype)initWithDictionary:(NSDictionary *)dict;
 /// 将 config 转化成字典
 -(NSDictionary *)convertToDictionary;
-@end
 
-@interface FTLoggerConfig ()
-/// 私有的初始化方法，通过字典来初始化，用于 Extensin SDK
-/// - Parameter dict: config 转化后的字典
--(instancetype)initWithDictionary:(NSDictionary *)dict;
-/// 将 config 转化成字典
--(NSDictionary *)convertToDictionary;
+/// 合并 remoteConfig
+-(void)mergeWithRemoteConfigDict:(NSDictionary *)dict;
 @end
 
 @interface FTRumConfig ()
@@ -38,6 +33,8 @@ NS_ASSUME_NONNULL_BEGIN
 -(instancetype)initWithDictionary:(NSDictionary *)dict;
 /// 将 config 转化成字典
 -(NSDictionary *)convertToDictionary;
+/// 合并 remoteConfig
+-(void)mergeWithRemoteConfigDict:(NSDictionary *)dict;
 @end
 
 @interface FTTraceConfig ()
@@ -46,5 +43,7 @@ NS_ASSUME_NONNULL_BEGIN
 -(instancetype)initWithDictionary:(NSDictionary *)dict;
 /// 将 config 转化成字典
 -(NSDictionary *)convertToDictionary;
+/// 合并 remoteConfig
+-(void)mergeWithRemoteConfigDict:(NSDictionary *)dict;
 @end
 NS_ASSUME_NONNULL_END