+++ FTMobileSDK/FTMobileAgent/Core/FTMobileAgent.m
@@ -25,9 +25,6 @@
 #import "FTUserInfo.h"
 #import "FTExtensionDataManager.h"
 #import "FTExternalDataManager+Private.h"
-#if !TARGET_OS_TV
-#import "FTWKWebViewHandler.h"
-#endif
 #import "FTMobileAgentVersion.h"
 #import "FTNetworkInfoManager.h"
 #import "FTURLSessionInstrumentation.h"
@@ -37,7 +34,8 @@
 #import "NSDictionary+FTCopyProperties.h"
 #import "FTTrackerEventDBTool.h"
 #import "FTDataWriterWorker.h"
-@interface FTMobileAgent ()<FTAppLifeCycleDelegate>
+#import "FTRemoteConfigManager.h"
+@interface FTMobileAgent ()<FTAppLifeCycleDelegate,FTRemoteConfigurationProtocol>
 @property (nonatomic, strong) FTLoggerConfig *loggerConfig;
 @property (nonatomic, strong) FTRumConfig *rumConfig;
 @property (nonatomic, strong) FTTraceConfig *traceConfig;
@@ -66,35 +64,12 @@ static dispatch_once_t onceToken;
         self = [super init];
         if (self) {
             _sdkConfig = [config copy];
-            //基础类型的记录
-            [FTLog enableLog:_sdkConfig.enableSDKDebugLog];
-            NSString *version = [[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleShortVersionString"];
-            [[FTPresetProperty sharedInstance] setDataModifier:_sdkConfig.dataModifier lineDataModifier:_sdkConfig.lineDataModifier];
-            [[FTPresetProperty sharedInstance] startWithVersion:version
-                                                     sdkVersion:SDK_VERSION
-                                                            env:_sdkConfig.env
-                                                        service:_sdkConfig.service
-                                                  globalContext:_sdkConfig.globalContext
-                                                        pkgInfo:_sdkConfig.pkgInfo
-            ];
-            [FTExtensionDataManager sharedInstance].groupIdentifierArray = _sdkConfig.groupIdentifiers;
-            //开启数据处理管理器
-            [FTTrackDataManager startWithAutoSync:_sdkConfig.autoSync syncPageSize:_sdkConfig.syncPageSize syncSleepTime:_sdkConfig.syncSleepTime];
-            if(config.enableLimitWithDbSize){
-                [[FTTrackDataManager sharedInstance] setDBLimitWithSize:_sdkConfig.dbCacheLimit discardNew:_sdkConfig.dbDiscardType == FTDBDiscard];
+            if (_sdkConfig.remoteConfiguration) {
+                [[FTRemoteConfigManager sharedManager] enable:YES updateInterval:_sdkConfig.remoteConfigMiniUpdateInterval];
+                [FTRemoteConfigManager sharedManager].delegate = self;
+                [_sdkConfig mergeWithRemoteConfigDict:[[FTRemoteConfigManager sharedManager] getLocalRemoteConfig]];
             }
-            
-            [FTNetworkInfoManager sharedInstance]
-                .setDatakitUrl(_sdkConfig.datakitUrl)
-                .setDatawayUrl(_sdkConfig.datawayUrl)
-                .setClientToken(_sdkConfig.clientToken)
-                .setSdkVersion(SDK_VERSION)
-                .setCompression(_sdkConfig.compressIntakeRequests)
-                .setEnableDataIntegerCompatible(_sdkConfig.enableDataIntegerCompatible);
-            [[FTURLSessionInstrumentation sharedInstance] setSdkUrlStr:_sdkConfig.datakitUrl.length>0?_sdkConfig.datakitUrl:_sdkConfig.datawayUrl
-                                                           serviceName:_sdkConfig.service];
-            [[FTExtensionDataManager sharedInstance] writeMobileConfig:[_sdkConfig convertToDictionary]];
-            FTInnerLogInfo(@"Init Mobile Config Success: \n%@",_sdkConfig.debugDescription);
+            [self applyBaseConfig:_sdkConfig];
         }
     }@catch(NSException *exception) {
         FTInnerLogError(@"exception: %@",exception);
@@ -105,21 +80,13 @@ static dispatch_once_t onceToken;
     NSAssert((rumConfigOptions.appid.length!=0 ), @"请设置 appid 用户访问监测应用ID");
     @try {
         if(!_rumConfig){
-            FTInnerLogInfo(@"[RUM] APPID:%@",rumConfigOptions.appid);
             _rumConfig = [rumConfigOptions copy];
-            [[FTPresetProperty sharedInstance] setRUMAppID:_rumConfig.appid sampleRate:_rumConfig.samplerate sessionOnErrorSampleRate:_rumConfig.sessionOnErrorSampleRate rumGlobalContext:_rumConfig.globalContext];
-            [[FTTrackDataManager sharedInstance] setRUMCacheLimitCount:_rumConfig.rumCacheLimitCount discardNew:_rumConfig.rumDiscardType == FTRUMDiscard];
-            [[FTGlobalRumManager sharedInstance] setRumConfig:_rumConfig writer:[FTTrackDataManager sharedInstance].dataWriterWorker];
-            [[FTURLSessionInstrumentation sharedInstance]setEnableAutoRumTrace:_rumConfig.enableTraceUserResource
-                                                            resourceUrlHandler:_rumConfig.resourceUrlHandler
-                                                      resourcePropertyProvider:_rumConfig.resourcePropertyProvider];
-            [[FTURLSessionInstrumentation sharedInstance] setRumResourceHandler:[FTGlobalRumManager sharedInstance].rumManager];
-            [FTExternalDataManager sharedManager].resourceDelegate = [FTURLSessionInstrumentation sharedInstance].externalResourceHandler;
-            [[FTExtensionDataManager sharedInstance] writeRumConfig:[_rumConfig convertToDictionary]];
-            if (_loggerConfig) {
-                [FTLogger sharedInstance].linkRumDataProvider = [FTGlobalRumManager sharedInstance].rumManager;
+            [FTNetworkInfoManager sharedInstance].setAppId(_rumConfig.appid);
+            if (_sdkConfig.remoteConfiguration) {
+                [[FTRemoteConfigManager sharedManager] updateRemoteConfig];
+                [_rumConfig mergeWithRemoteConfigDict:[[FTRemoteConfigManager sharedManager] getLocalRemoteConfig]];
             }
-            FTInnerLogInfo(@"Init RUM Config Success: \n%@",_rumConfig.debugDescription);
+            [self applyRUMConfig:_rumConfig];
         }
     } @catch (NSException *exception) {
         FTInnerLogError(@"exception: %@",exception);
@@ -127,17 +94,13 @@ static dispatch_once_t onceToken;
 }
 - (void)startLoggerWithConfigOptions:(FTLoggerConfig *)loggerConfigOptions{
     @try {
-        
         if (!_loggerConfig) {
             _loggerConfig = [loggerConfigOptions copy];
-            [[FTPresetProperty sharedInstance] setLogGlobalContext:_loggerConfig.globalContext];
-            [[FTTrackDataManager sharedInstance] setLogCacheLimitCount:_loggerConfig.logCacheLimitCount discardNew:_loggerConfig.discardType == FTDiscard];
-            [[FTExtensionDataManager sharedInstance] writeLoggerConfig:[_loggerConfig convertToDictionary]];
-            [FTLogger startWithLoggerConfig:_loggerConfig writer:[FTTrackDataManager sharedInstance].dataWriterWorker];
-            [FTLogger sharedInstance].linkRumDataProvider = [FTGlobalRumManager sharedInstance].rumManager;
-            FTInnerLogInfo(@"Init Logger Config Success: \n%@",_loggerConfig.debugDescription);
+            if (_sdkConfig.remoteConfiguration) {
+                [_loggerConfig mergeWithRemoteConfigDict:[[FTRemoteConfigManager sharedManager] getLocalRemoteConfig]];
+            }
+            [self applyLogConfig:_loggerConfig];
         }
-        
     } @catch (NSException *exception) {
         FTInnerLogError(@"exception: %@",exception);
     }
@@ -146,25 +109,99 @@ static dispatch_once_t onceToken;
     @try {
         if(!_traceConfig){
             _traceConfig = [traceConfigOptions copy];
-#if !TARGET_OS_TV
-            [FTWKWebViewHandler sharedInstance].enableTrace = _traceConfig.enableAutoTrace;
-            [FTWKWebViewHandler sharedInstance].interceptor = [FTURLSessionInstrumentation sharedInstance].interceptor;
-#endif
-            [[FTURLSessionInstrumentation sharedInstance] setTraceEnableAutoTrace:_traceConfig.enableAutoTrace
-                                                                enableLinkRumData:_traceConfig.enableLinkRumData
-                                                                       sampleRate:_traceConfig.samplerate
-                                                                        traceType:_traceConfig.networkTraceType
-                                                                 traceInterceptor:_traceConfig.traceInterceptor
-            ];
-            [FTExternalDataManager sharedManager].resourceDelegate = [FTURLSessionInstrumentation sharedInstance].externalResourceHandler;
-            [[FTExtensionDataManager sharedInstance] writeTraceConfig:[_traceConfig convertToDictionary]];
-            FTInnerLogInfo(@"Init Trace Config Success: \n%@",_traceConfig.debugDescription);
+            if (_sdkConfig.remoteConfiguration) {
+                [_traceConfig mergeWithRemoteConfigDict:[[FTRemoteConfigManager sharedManager] getLocalRemoteConfig]];
+            }else{
+                [self applyTraceConfig:_traceConfig];
+            }
         }
     } @catch (NSException *exception) {
         FTInnerLogError(@"exception: %@",exception);
     }
-    
 }
+#pragma mark ========== remote ==========
+- (void)updateRemoteConfiguration:(NSDictionary *)configuration{
+    [[FTTrackDataManager sharedInstance] updateWithRemoteConfiguration:configuration];
+    [[FTLogger sharedInstance] updateWithRemoteConfiguration:configuration];
+}
++ (void)updateRemoteConfig{
+    if (onceToken == 0 && sharedInstance == nil) {
+        return;
+    }
+    [[FTRemoteConfigManager sharedManager] updateRemoteConfig];
+}
++ (void)updateRemoteConfigWithMiniUpdateInterval:(int)miniUpdateInterval callback:(void (^)(BOOL, NSDictionary<NSString *,id> * _Nullable))callback{
+    if (onceToken == 0 && sharedInstance == nil) {
+        return;
+    }
+    [[FTRemoteConfigManager sharedManager] updateRemoteConfigWithMiniUpdateInterval:miniUpdateInterval callback:callback];
+}
+#pragma mark ========== real sdk init ==========
+- (void)applyBaseConfig:(FTMobileConfig *)config{
+    //基础类型的记录
+    [FTLog enableLog:config.enableSDKDebugLog];
+    NSString *version = [[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleShortVersionString"];
+    [[FTPresetProperty sharedInstance] setDataModifier:config.dataModifier lineDataModifier:config.lineDataModifier];
+    [[FTPresetProperty sharedInstance] startWithVersion:version
+                                             sdkVersion:SDK_VERSION
+                                                    env:config.env
+                                                service:config.service
+                                          globalContext:config.globalContext
+                                                pkgInfo:config.pkgInfo
+    ];
+    [FTExtensionDataManager sharedInstance].groupIdentifierArray = config.groupIdentifiers;
+    //开启数据处理管理器
+    [FTTrackDataManager startWithAutoSync:config.autoSync syncPageSize:config.syncPageSize syncSleepTime:config.syncSleepTime];
+    [[FTTrackDataManager sharedInstance] setEnableLimitWithDb:config.enableLimitWithDbSize size:config.dbCacheLimit discardNew:config.dbDiscardType == FTDBDiscard];
+    
+    [FTNetworkInfoManager sharedInstance]
+        .setDatakitUrl(config.datakitUrl)
+        .setDatawayUrl(config.datawayUrl)
+        .setClientToken(config.clientToken)
+        .setSdkVersion(SDK_VERSION)
+        .setCompression(config.compressIntakeRequests)
+        .setEnableDataIntegerCompatible(config.enableDataIntegerCompatible);
+    [[FTURLSessionInstrumentation sharedInstance] setSdkUrlStr:config.datakitUrl.length>0?config.datakitUrl:config.datawayUrl
+                                                   serviceName:config.service];
+    [[FTExtensionDataManager sharedInstance] writeMobileConfig:[config convertToDictionary]];
+    FTInnerLogInfo(@"Init Mobile Config Success: \n%@",config.debugDescription);
+}
+- (void)applyRUMConfig:(FTRumConfig *)rumConfig{
+    FTInnerLogInfo(@"[RUM] APPID:%@",rumConfig.appid);
+    [[FTPresetProperty sharedInstance] setRUMAppID:rumConfig.appid sampleRate:rumConfig.samplerate sessionOnErrorSampleRate:rumConfig.sessionOnErrorSampleRate rumGlobalContext:rumConfig.globalContext];
+    [[FTTrackDataManager sharedInstance] setRUMCacheLimitCount:rumConfig.rumCacheLimitCount discardNew:rumConfig.rumDiscardType == FTRUMDiscard];
+    [[FTGlobalRumManager sharedInstance] setRumConfig:rumConfig writer:[FTTrackDataManager sharedInstance].dataWriterWorker];
+    [[FTURLSessionInstrumentation sharedInstance]setEnableAutoRumTrace:rumConfig.enableTraceUserResource
+                                                    resourceUrlHandler:rumConfig.resourceUrlHandler
+                                              resourcePropertyProvider:rumConfig.resourcePropertyProvider];
+    [[FTURLSessionInstrumentation sharedInstance] setRumResourceHandler:[FTGlobalRumManager sharedInstance].rumManager];
+    [FTExternalDataManager sharedManager].resourceDelegate = [FTURLSessionInstrumentation sharedInstance].externalResourceHandler;
+    [[FTExtensionDataManager sharedInstance] writeRumConfig:[rumConfig convertToDictionary]];
+    if (_loggerConfig) {
+        [FTLogger sharedInstance].linkRumDataProvider = [FTGlobalRumManager sharedInstance].rumManager;
+    }
+    FTInnerLogInfo(@"Init RUM Config Success: \n%@",rumConfig.debugDescription);
+}
+- (void)applyLogConfig:(FTLoggerConfig *)loggerConfig{
+    [[FTPresetProperty sharedInstance] setLogGlobalContext:loggerConfig.globalContext];
+    [[FTTrackDataManager sharedInstance] setLogCacheLimitCount:loggerConfig.logCacheLimitCount discardNew:loggerConfig.discardType == FTDiscard];
+    [[FTExtensionDataManager sharedInstance] writeLoggerConfig:[loggerConfig convertToDictionary]];
+    [FTLogger startWithLoggerConfig:loggerConfig writer:[FTTrackDataManager sharedInstance].dataWriterWorker];
+    [FTLogger sharedInstance].linkRumDataProvider = [FTGlobalRumManager sharedInstance].rumManager;
+    FTInnerLogInfo(@"Init Logger Config Success: \n%@",loggerConfig.debugDescription);
+}
+- (void)applyTraceConfig:(FTTraceConfig *)traceConfig{
+    [[FTURLSessionInstrumentation sharedInstance] setTraceEnableAutoTrace:traceConfig.enableAutoTrace
+                                                        enableLinkRumData:traceConfig.enableLinkRumData
+                                                               sampleRate:traceConfig.samplerate
+                                                                traceType:traceConfig.networkTraceType
+                                                         traceInterceptor:traceConfig.traceInterceptor
+    ];
+    [FTExternalDataManager sharedManager].resourceDelegate = [FTURLSessionInstrumentation sharedInstance].externalResourceHandler;
+    [[FTExtensionDataManager sharedInstance] writeTraceConfig:[traceConfig convertToDictionary]];
+    FTInnerLogInfo(@"Init Trace Config Success: \n%@",traceConfig.debugDescription);
+}
+
 #pragma mark ========== public method ==========
 - (void)isIntakeUrl:(BOOL(^)(NSURL *url))handler{
     if(handler){