+++ FTMobileSDK/FTSDKCore/Logger/FTLogger.m
@@ -16,8 +16,8 @@
 #import "NSDictionary+FTCopyProperties.h"
 #import "FTEnumConstant.h"
 #import "FTPresetProperty.h"
-#import "FTLoggerConfig.h"
-
+#import "FTLoggerConfig+Private.h"
+#import "FTJSONUtil.h"
 void *FTLoggerQueueIdentityKey = &FTLoggerQueueIdentityKey;
 
 @interface FTLogger ()
@@ -45,52 +45,47 @@ static dispatch_once_t onceToken;
     if(self){
         _config = config;
         _loggerWriter = writer;
-        if (config.logLevelFilter) {
-            _logLevelFilterSet = [NSSet setWithArray:config.logLevelFilter];
-        }
         _loggerQueue = dispatch_queue_create("com.guance.logger", DISPATCH_QUEUE_SERIAL);
         dispatch_queue_set_specific(_loggerQueue,FTLoggerQueueIdentityKey, &FTLoggerQueueIdentityKey, NULL);
+        [self dealLogLevelFilter:config.logLevelFilter];
     }
     return self;
+}
+-(void)dealLogLevelFilter:(NSArray *)logLevelFilter{
+    if (!logLevelFilter || logLevelFilter.count == 0) {
+        return;
+    }
+    NSMutableArray *levels = [[NSMutableArray alloc]init];
+    for (id level in logLevelFilter) {
+        if ([level isKindOfClass:NSNumber.class] && [level intValue]<5 && [level intValue]>=0) {
+            [levels addObject:FTStatusStringMap[[level intValue]]];
+        }else{
+            [levels addObject:level];
+        }
+    }
+    _logLevelFilterSet = [NSSet setWithArray:levels];
+}
+-(void)updateWithRemoteConfiguration:(NSDictionary *)configuration{
+    __weak typeof(self) weakSelf = self;
+    dispatch_async(self.loggerQueue, ^{
+        __strong typeof(weakSelf) strongSelf = weakSelf;
+        if (!strongSelf) return;
+        [strongSelf.config mergeWithRemoteConfigDict:configuration];
+        [strongSelf dealLogLevelFilter:strongSelf.config.logLevelFilter];
+    });
 }
 - (void)log:(NSString *)content
      statusType:(FTLogStatus)statusType
    property:(nullable NSDictionary *)property{
     NSDictionary *copyDict = [property ft_deepCopy];
-    if (!content || content.length == 0 ) {
-        FTInnerLogError(@"[Logging] 传入的第数据格式有误");
-        return;
-    }
-    if(self.config.printCustomLogToConsole){
-        FT_CONSOLE_LOG((LogStatus)statusType,content,copyDict);
-    }
-    if (!self.config.enableCustomLog) {
-        return;
-    }
-    if (self.logLevelFilterSet && ![self.logLevelFilterSet containsObject:@(statusType)]) {
-        return;
-    }
-    [self _log:content status:FTStatusStringMap[statusType] property:copyDict];
+    [self _log:content statusType:(LogStatus)statusType status:FTStatusStringMap[statusType] property:copyDict];
 }
 - (void)log:(NSString *)content status:(NSString *)status{
     [self log:content status:status property:nil];
 }
 - (void)log:(NSString *)content status:(NSString *)status property:(nullable NSDictionary *)property{
     NSDictionary *copyDict = [property ft_deepCopy];
-    if (!content || content.length == 0 ) {
-        FTInnerLogError(@"[Logging] 传入的第数据格式有误");
-        return;
-    }
-    if(self.config.printCustomLogToConsole){
-        FT_CONSOLE_CUSTOM_LOG(status, content, copyDict);
-    }
-    if (!self.config.enableCustomLog) {
-        return;
-    }
-    if (self.logLevelFilterSet && ![self.logLevelFilterSet containsObject:status]) {
-        return;
-    }
-    [self _log:content status:status property:copyDict];
+    [self _log:content statusType:StatusCustom status:status property:copyDict];
 }
 -(void)info:(NSString *)content property:(NSDictionary *)property{
     [self log:content statusType:FTStatusInfo property:property];
@@ -107,7 +102,21 @@ static dispatch_once_t onceToken;
 - (void)ok:(NSString *)content property:(NSDictionary *)property{
     [self log:content statusType:FTStatusOk property:property];
 }
-- (void)_log:(NSString *)content status:(id)status property:(nullable NSDictionary *)property{
+- (void)_log:(NSString *)content statusType:(LogStatus)type status:(NSString *)status property:(nullable NSDictionary *)property{
+    if (!content || content.length == 0 ) {
+        FTInnerLogError(@"[Logging] 传入的第数据格式有误");
+        return;
+    }
+    NSDictionary *copyDict = [property ft_deepCopy];
+    if(self.config.printCustomLogToConsole){
+        FT_CONSOLE_LOG(type,status,content,copyDict);
+    }
+    if (!self.config.enableCustomLog) {
+        return;
+    }
+    if (self.logLevelFilterSet && ![self.logLevelFilterSet containsObject:status]) {
+        return;
+    }
     if (![FTBaseInfoHandler randomSampling:self.config.samplerate]){
         FTInnerLogInfo(@"[Logging][Not Sampled] %@",content);
         return;